requirements.txt
streamlit
pandas
python-dotenv
web3
requests
urllib3
